moved code for getting excel to common area and using it in schedule and report controller
